requirements.txt
# Core dependencies
dspy-ai>=2.4.0
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data and I/O
jsonlines>=3.1.0
tqdm>=4.65.0
rich>=13.0.0

# Optional backends
anthropic>=0.25.0
ollama>=0.2.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=22.0.0
isort>=5.0.0
ruff>=0.1.0
